requirements.txt
ladybug-core>=0.41.6
